+++ modules/bim/lib/open_project/bim/engine.rb
@@ -207,7 +207,7 @@ module OpenProject::Bim
 
     config.to_prepare do
       ::WorkPackage::Exporter
-        .register_for_list(:bcf, OpenProject::Bim::BcfXml::DelayedExporter)
+        .register_for_list(:bcf, OpenProject::Bim::BcfXml::Exporter)
 
       ::Queries::Register.filter ::Query, OpenProject::Bim::BcfIssueAssociatedFilter
       ::Queries::Register.column ::Query, OpenProject::Bim::QueryBcfThumbnailColumn
