--- modules/xls_export/spec/lib/work_package_xls_export_spec.rb
@@ -1,171 +1,0 @@
-require 'spec_helper'
-require 'spreadsheet'
-
-describe "WorkPackageXlsExport" do
-  let(:project) { FactoryBot.create :project }
-
-  let(:parent) { FactoryBot.create :work_package, project: project, subject: 'Parent' }
-  let(:child_1) do
-    FactoryBot.create :work_package, parent: parent, project: project, subject: 'Child 1'
-  end
-  let(:child_2) do
-    FactoryBot.create :work_package, parent: parent, project: project, subject: 'Child 2'
-  end
-
-  let(:single) { FactoryBot.create :work_package, project: project, subject: 'Single' }
-  let(:followed) { FactoryBot.create :work_package, project: project, subject: 'Followed' }
-
-  let(:child_2_child) do
-    FactoryBot.create :work_package, parent: child_2, project: project, subject: "Child 2's child"
-  end
-
-  let(:relation) do
-    child_2.new_relation.tap do |r|
-      r.to = followed
-      r.relation_type = 'follows'
-      r.delay = 0
-      r.description = 'description foobar'
-      r.save
-    end
-  end
-
-  let(:work_packages) do
-    work_packages = [parent, child_1, child_2, single, followed]
-    child_2_child
-
-    work_packages
-  end
-
-  let(:current_user) { FactoryBot.create :admin } # may export relations
-
-  let(:query) do
-    query = FactoryBot.build(:query, user: current_user, project: project)
-
-    query.filters.clear
-    query.sort_criteria = [['id', 'asc']]
-    query
-  end
-
-  let(:sheet) do
-    login_as(current_user)
-    work_packages
-    relation
-    work_packages.each(&:reload) # to init .leaves and relations
-    load_sheet export
-  end
-
-  let(:export) do
-    OpenProject::XlsExport::WorkPackageXlsExport.new(
-      query,
-      show_relations: true
-    )
-  end
-
-  def load_sheet(export)
-    f = Tempfile.new 'result.xls'
-    begin
-      f.binmode
-      f.write export.list(&:content)
-    ensure
-      f.close
-    end
-
-    sheet = Spreadsheet.open(f.path).worksheets.first
-    f.unlink
-
-    sheet
-  end
-
-  # expected row and column indices
-
-  PARENT = 2
-  CHILD_1 = 4
-  CHILD_2 = 5
-  SINGLE = 8
-  FOLLOWED = 9
-  RELATION = 8
-  RELATION_DESCRIPTION = 10
-  RELATED_SUBJECT = 13
-
-  it 'produces the correct result' do
-    expect(query.columns.map(&:name)).to eq [:type, :id, :subject, :status, :assigned_to, :priority]
-
-    # the first header row devides the sheet into work packages and relation columns
-    expect(sheet.rows.first.take(8)).to eq ['Work packages', nil, nil, nil, nil, nil, nil, 'Relations']
-
-    # the second header row includes the column names for work packages and relations and the related work package
-    expect(sheet.rows[1])
-      .to eq [
-        nil, 'Type', 'ID', 'Subject', 'Status', 'Assignee', 'Priority',
-        nil, 'Relation type', 'Delay', 'Description',
-        'Type', 'ID', 'Subject', 'Status', 'Assignee', 'Priority',
-        nil
-      ]
-
-    # duplicates rows for each relation
-    c2id = child_2.id
-    expect(sheet.column(2).drop(2))
-      .to eq [parent.id, parent.id, child_1.id, c2id, c2id, c2id, single.id, followed.id, child_2_child.id]
-
-    # marks Parent as parent of Child 1 and 2
-    expect(sheet.row(PARENT)[RELATION]).to eq 'parent of'
-    expect(sheet.row(PARENT)[RELATED_SUBJECT]).to eq 'Child 1'
-
-    expect(sheet.row(PARENT + 1)[RELATION]).to eq 'parent of'
-    expect(sheet.row(PARENT + 1)[RELATED_SUBJECT]).to eq 'Child 2'
-
-    # shows Child 1 as child of Parent
-    expect(sheet.row(CHILD_1)[RELATION]).to eq 'child of'
-    expect(sheet.row(CHILD_1)[RELATED_SUBJECT]).to eq 'Parent'
-
-    # shows Child 2 as child of Parent
-    expect(sheet.row(CHILD_2)[RELATION]).to eq 'child of'
-    expect(sheet.row(CHILD_2)[RELATED_SUBJECT]).to eq 'Parent'
-
-    # shows Child 2 as parent of Child 2's child
-    expect(sheet.row(CHILD_2 + 1)[RELATION]).to eq 'parent of'
-    expect(sheet.row(CHILD_2 + 1)[RELATED_SUBJECT]).to eq "Child 2's child"
-
-    # shows Child 2 as following Followed
-    expect(sheet.row(CHILD_2 + 2)[RELATION]).to eq 'Follows'
-    expect(sheet.row(CHILD_2 + 2)[RELATED_SUBJECT]).to eq 'Followed'
-
-    # shows no relation information for Single
-    expect(sheet.row(SINGLE).drop(7).compact).to eq []
-
-    # shows Followed as preceding Child 2'
-    expect(sheet.row(FOLLOWED)[RELATION]).to eq 'Precedes'
-    expect(sheet.row(FOLLOWED)[RELATION_DESCRIPTION]).to eq 'description foobar'
-    expect(sheet.row(FOLLOWED)[RELATED_SUBJECT]).to eq 'Child 2'
-
-    # exports the correct data (examples)
-    expect(sheet.row(PARENT))
-      .to eq [
-        nil, parent.type.name, parent.id, parent.subject, parent.status.name, parent.assigned_to, parent.priority.name,
-        nil, 'parent of', nil, nil,
-        child_1.type.name, child_1.id, child_1.subject, child_1.status.name, child_1.assigned_to, child_1.priority.name
-      ] # delay nil as this is a parent-child relation not represented by an actual Relation record
-
-    expect(sheet.row(SINGLE))
-      .to eq [
-        nil, single.type.name, single.id, single.subject, single.status.name, single.assigned_to, single.priority.name
-      ]
-
-    expect(sheet.row(FOLLOWED))
-      .to eq [
-        nil,
-        followed.type.name, followed.id, followed.subject, followed.status.name, followed.assigned_to, followed.priority.name,
-        nil, 'Precedes', 0, relation.description,
-        child_2.type.name, child_2.id, child_2.subject, child_2.status.name, child_2.assigned_to, child_2.priority.name
-      ]
-  end
-
-  context 'with someone who may not see related work packages' do
-    let(:current_user) { FactoryBot.create :user }
-
-    it 'exports no information without visibility' do
-      expect(sheet.rows.length).to eq(2)
-      expect(sheet.column(1).drop(2)).to be_empty
-    end
-  end
-end
